hooks/utils: move utility functions from _tkinter hook to separate module
Move utility functions from hook-_tkinter.py into separate module,
utils.hooks.tcl_tk. This will allow reuse of the collect_tcl_tk_files()
function in the Tcl/Tk-based splash screen support code (#4887).

The idea is to perform this move separately and in controlled
conditions, thereby ensuring that we lose none of the bugfixes
that were added to the _tkinter hook after #4887 was openend.

It will also simplify the rebase of #4887, which will require
only dropping of the corresponding original commit and minor
adjustments in the subsequent commits.
